Shift in the destination width to silence MSVC C4334
The first native Windows build emitted three C4334 warnings ("result of
32-bit shift implicitly converted to 64 bits"): two in bench_main.cu and
one in pipeline.cuh.

All three are benign. logI is bounded to [2,20] and log_region to [1,30]
before any of these run -- there is already a comment at the log_region
bound saying it exists so that no `1u << log_region` is ever undefined --
so a 32-bit shift could not overflow and the widening afterwards loses
nothing. C4334 fires on the pattern, not on the proof.

Changed anyway, because the Linux build is warning-free and a Windows
build that always emits three warnings is a Windows build whose warnings
stop being read. `(uint64_t)1 << n` also says what was meant.

Verified: no behaviour change. A full band (q 50000000-50001999, logI 16,
J 16384, 5574 relations) hashes identically before and after, and the
build is clean on both platforms.

+++ bench/bench_main.cu
@@ -1234,6 +1234,12 @@ static int bench_main_impl(int argc, char **argv)
     if (!fbbound && cfg.side == 1 && !cfg.pipeline)
         fbbound = (q > 0xFFFFFFFFull) ? 0xFFFFFFFFu : (uint32_t)q;
 
+    /* Shift in the destination width, not the source width. These are all
+     * provably in range -- logI is bounded to [2,20] and log_region to [1,30]
+     * above, so a 32-bit shift could not overflow -- but MSVC's C4334 fires on
+     * the pattern rather than on the proof, and a Windows build that always
+     * emits three warnings is a Windows build whose warnings get ignored.
+     * `(uint64_t)1 << n` says what was meant and costs nothing. */
     /* Bound log_region before ANY 1u << log_region: the shift is undefined for
      * >= 32 and UBSan flags --region 32 on the old ordering. */
     if (cfg.log_region < 1 || cfg.log_region > 30) {
@@ -1251,7 +1257,7 @@ static int bench_main_impl(int argc, char **argv)
             return 1;
         }
     }
-    if ((uint64_t)(1u << cfg.logI) * cfg.J > 0x80000000ull) {
+    if (((uint64_t)1 << cfg.logI) * cfg.J > 0x80000000ull) {
         fprintf(stderr, "I*J must fit in 31 bits (uint32 positions)\n"); return 1;
     }
     /* Two limits that used to be silent. A 2 B or 4 B record carries the
@@ -1321,7 +1327,7 @@ static int bench_main_impl(int argc, char **argv)
                 " hangs and a partial warp double-counts\n");
         return 1;
     }
-    if ((uint64_t)(1u << cfg.logI) * cfg.J % (1u << cfg.log_region)) {
+    if (((uint64_t)1 << cfg.logI) * cfg.J % ((uint64_t)1 << cfg.log_region)) {
         fprintf(stderr, "I*J must divide evenly into 2^%d regions\n", cfg.log_region);
         return 1;
     }
